requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
pandas==2.1.4
pyyaml==6.0.1
openai==1.3.7
python-dotenv==1.0.0
feedparser==6.0.10